[REF] Core: handle toggleList in List plugin
