+++ packages/plugin-list/List.ts
@@ -3,6 +3,9 @@ import { ParsingFunction, ParsingContext, ParsingMap } from '../core/src/Parser'
 import { ListNode, ListType } from './ListNode';
 import { ParagraphNode } from '../plugin-paragraph/ParagraphNode';
 import { utils } from '../utils/src/utils';
+import { ListParams } from '../core/src/CorePlugin';
+import { VNode } from '../core/src/VNodes/VNode';
+import { withMarkers } from '../utils/src/markers';
 
 const listTags = ['UL', 'OL'];
 const listContextTags = listTags.concat('LI');
@@ -16,6 +19,11 @@ function _isEmptyTextNode(node: Node): boolean {
 }
 
 export class List extends JWPlugin {
+    commands = {
+        toggleList: {
+            handler: this.toggleList.bind(this),
+        },
+    };
     static readonly parsingFunctions: Array<ParsingFunction> = [List.parse];
     static parse(context: ParsingContext): [ParsingContext, ParsingMap] {
         if (listTags.includes(context.currentNode.nodeName)) {
@@ -79,5 +87,188 @@ export class List extends JWPlugin {
         // the paragraph created above, or to the list itself in the case of
         // blocks.
         return [context, parsingMap];
+    }
+
+    //--------------------------------------------------------------------------
+    // Public
+    //--------------------------------------------------------------------------
+
+    /**
+     * Insert/remove a list at range.
+     *
+     * @param params
+     */
+    toggleList(params: ListParams): void {
+        const type = params.type;
+        const range = params.range || this.editor.vDocument.selection.range;
+
+        // Retrieve the nodes in the selection.
+        let selectedNodes: Array<VNode>;
+        if (range.isCollapsed()) {
+            // Toggle the parent if the range is collapsed.
+            selectedNodes = [range.start.parent];
+        } else {
+            selectedNodes = range.selectedNodes();
+        }
+
+        // Check if all selected nodes are within a list of the same type.
+        const areAllInMatchingList = selectedNodes.every(node => {
+            const listAncestor = node.ancestor(ListNode);
+            return listAncestor && listAncestor.listType === type;
+        });
+
+        // Dispatch.
+        if (areAllInMatchingList) {
+            // If all selected nodes are within a list of the same type,
+            // "unlist" them.
+            this._unlist(selectedNodes);
+        } else {
+            this._convertToList(selectedNodes, type);
+        }
+    }
+
+    //--------------------------------------------------------------------------
+    // Private
+    //--------------------------------------------------------------------------
+
+    /**
+     * Convert the given nodes to a list with the given `htmlTag`.
+     *
+     * @param nodes
+     * @param type
+     */
+    _convertToList(nodes: Array<VNode>, type: ListType): void {
+        let newList = new ListNode(type);
+        let duplicatedNodes = [];
+
+        // Lists cannot have atomic nodes as direct children.
+        nodes = Array.from(new Set(nodes.map(node => (node.atomic ? node.parent : node))));
+
+        // If the last node is within a list, we need to split that list.
+        const last = nodes[nodes.length - 1];
+        const lastListAncestor = last.ancestor(ListNode);
+        if (lastListAncestor) {
+            const duplicatedAncestors = this._splitUpToAncestor(last, lastListAncestor);
+            duplicatedNodes = [...duplicatedNodes, ...duplicatedAncestors];
+        }
+
+        // If the first node is within a list, we need to split that list.
+        // Then insert the new list to which the nodes will be appended.
+        const first = nodes[0];
+        const firstListAncestor = first.ancestor(ListNode);
+        if (firstListAncestor) {
+            const duplicatedAncestors = this._splitUpToAncestor(first, firstListAncestor);
+            duplicatedNodes = [...duplicatedNodes, ...duplicatedAncestors];
+
+            // Insert the new list before the first node's list ancestor.
+            // It was split so we need its duplicate as that is where the node is.
+            const newListAncestor = duplicatedAncestors[duplicatedAncestors.length - 1];
+            newListAncestor.before(newList);
+        } else {
+            // Insert the new list before the common ancestor to the first and
+            // last nodes.
+            const commonAncestor = first.commonAncestor(last);
+            const reference = first.ancestor(node => node.parent === commonAncestor) || first;
+            commonAncestor.insertBefore(newList, reference);
+        }
+
+        // Move the nodes to the list
+        withMarkers(() => {
+            // Remove the children of nested lists (they will be moved together
+            // with their parent) and convert nested lists to the new type.
+            const nestedLists = nodes.filter(this._isNestedList.bind(this));
+            const nonNestedNodes = nodes.filter(node => {
+                return !node.ancestor(ancestor => nestedLists.includes(ancestor));
+            });
+            nestedLists.forEach(nestedList => ((nestedList as ListNode).listType = type));
+            // Move the nodes.
+            nonNestedNodes.forEach(node => {
+                newList.append(node);
+            });
+            // Remove empty lists.
+            duplicatedNodes.concat(nodes).forEach(node => {
+                if (node.is(ListNode) && !node.hasChildren()) {
+                    node.remove();
+                }
+            });
+        });
+
+        // If the new list is after or before a list of the same type, merge
+        // them (eg: <ol><li>a</li></ol><ol><li>b</li></ol> =>
+        // <ol><li>a</li><li>b</li></ol>).
+        const previousSibling = newList.previousSibling();
+        if (previousSibling && previousSibling.is(ListNode) && previousSibling.listType === type) {
+            newList.children.slice().forEach(child => previousSibling.append(child));
+            newList.remove();
+            newList = previousSibling as ListNode;
+        }
+        const nextSibling = newList.nextSibling();
+        if (nextSibling && nextSibling.is(ListNode) && nextSibling.listType === type) {
+            nextSibling.children.slice().forEach(child => newList.append(child));
+            nextSibling.remove();
+        }
+    }
+    /**
+     * Return true if the given node is a nested `ListNode`, false otherwise.
+     *
+     * @param node
+     */
+    _isNestedList(node: VNode): node is ListNode {
+        return node.is(ListNode) && !!node.ancestor(ListNode);
+    }
+    /**
+     * Split a node and its ancestors up until the given ancestor.
+     *
+     * @param node
+     * @returns a list of the duplicated nodes generated during the splits.
+     */
+    _splitUpToAncestor(node: VNode, ancestor: VNode): VNode[] {
+        const duplicatedNodes = [];
+        let child = node;
+        let parent = child.parent;
+        do {
+            duplicatedNodes.push(parent.splitAt(child));
+            child = parent;
+            parent = child.parent;
+        } while (child !== ancestor);
+        if (!ancestor.hasChildren()) {
+            ancestor.remove();
+        }
+        return duplicatedNodes;
+    }
+    /**
+     * Turn list elements into non-list elements.
+     *
+     * @param nodes
+     */
+    _unlist(nodes: Array<VNode>): void {
+        // Get the direct children of each list to unlist.
+        const listItems = utils.distinct(
+            nodes.map(node => {
+                if (node.parent.is(ListNode)) {
+                    return node;
+                } else {
+                    return node.ancestor(ancestor => {
+                        return ancestor.parent.is(ListNode);
+                    });
+                }
+            }),
+        );
+
+        // Split the lists and move their contents to move out of them.
+        const lists = [];
+        listItems.forEach(item => {
+            lists.push(item.parent);
+            const newList = item.parent.splitAt(item);
+            newList.before(item);
+            lists.push(newList);
+        });
+
+        // Remove empty lists.
+        lists.forEach(list => {
+            if (!list.hasChildren()) {
+                list.remove();
+            }
+        });
     }
 }
